+++ src/pendule.cpp
@@ -12,7 +12,7 @@ Pendule::~Pendule(){
 
 }
 
-Pendule::Pendule(int id, long double m, long double r, long double theta0, long double omega0, Pendule *parent){
+Pendule::Pendule(int id, double m, double r, double theta0, double omega0, Pendule *parent){
     this->identifieur = id;
     this->masse = m;
     this->longeur = r;
@@ -22,7 +22,7 @@ Pendule::Pendule(int id, long double m, long double r, long double theta0, long 
     this->attache = parent;
 }
 
-Pendule::Pendule(int id, long double m, long double r, long double theta0, long double omega0){
+Pendule::Pendule(int id, double m, double r, double theta0, double omega0){
     this->identifieur = id;
     this->masse = m;
     this->longeur = r;
@@ -37,24 +37,24 @@ int Pendule::id() {
     return identifieur;
 }
 
-long double Pendule::r() {
+double Pendule::r() {
     return longeur;
 }
 
-long double Pendule::theta() {
+double Pendule::theta() {
     return th;
 }
 
-long double Pendule::oldomega() {
+double Pendule::oldomega() {
     return oldvit;
 }
 
-long double Pendule::omega() {
+double Pendule::omega() {
     return vit_angl;
 } 
 
-long double Pendule::x() {
-    long double tmp = longeur * sinl(th);
+double Pendule::x() {
+    double tmp = longeur * sin(th);
     if (attache == NULL) {
         return tmp;
     }
@@ -63,8 +63,8 @@ long double Pendule::x() {
     }
 }
 
-long double Pendule::y() {
-    long double tmp = - longeur * cosl(th);
+double Pendule::y() {
+    double tmp = - longeur * cos(th);
     if (attache == NULL) {
         return tmp;
     }
@@ -73,7 +73,7 @@ long double Pendule::y() {
     }
 }
 
-long double Pendule::m() {
+double Pendule::m() {
     return masse;
 }
 
@@ -87,23 +87,23 @@ void Pendule::id(int id) {
     this->identifieur = id;
 }
 
-void Pendule::r(long double r) {
+void Pendule::r(double r) {
     this->longeur = r;
 }
 
-void Pendule::theta(long double theta) {
+void Pendule::theta(double theta) {
     this->th = std::remainder(theta, 2 * M_PI);
 }
 
-void Pendule::oldomega(long double omega) {
+void Pendule::oldomega(double omega) {
     this->oldvit = omega;
 }
 
-void Pendule::omega(long double omega) {
+void Pendule::omega(double omega) {
     this->vit_angl = omega;
 }
 
-void Pendule::m(long double m) {
+void Pendule::m(double m) {
     this->masse = m;
 }
 
@@ -115,11 +115,11 @@ void Pendule::attacher(Pendule* parent) {
 
 void Pendule::print(){
     printf("id: %d\n", identifieur);
-    printf("\tmasse: %0.3Lg\n", masse);
-    printf("\tcoord cart: %0.3Lg, %0.3Lg\n", x(), y());
-    printf("\tcoord pol: %0.3Lg, %0.3Lg\n", longeur, th - M_PI_2);
-    printf("\tcoord nat: %0.3Lg, %0.3Lg\n", longeur, th);
-    printf("\tvitesse angulaire: %0.3Lg\n", vit_angl);
+    printf("\tmasse: %0.3g\n", masse);
+    printf("\tcoord cart: %0.3g, %0.3g\n", x(), y());
+    printf("\tcoord pol: %0.3g, %0.3g\n", longeur, th - M_PI_2);
+    printf("\tcoord nat: %0.3g, %0.3g\n", longeur, th);
+    printf("\tvitesse angulaire: %0.3g\n", vit_angl);
     if (attache != NULL){
         printf("\tid attache: %d\n", attache->id());
     } else {
